refactor: helper 클래스 적용
