+++ src/main/java/ssammudan/cotree/model/project/project/repository/ProjectRepositoryImpl.java
@@ -2,9 +2,7 @@ package ssammudan.cotree.model.project.project.repository;
 
 import java.util.Collections;
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
-import java.util.stream.Collectors;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
@@ -15,14 +13,13 @@ import com.querydsl.core.BooleanBuilder;
 import com.querydsl.jpa.impl.JPAQuery;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 
-import jakarta.persistence.EntityManager;
 import ssammudan.cotree.domain.project.dto.ProjectListResponse;
 import ssammudan.cotree.model.common.like.entity.QLike;
-import ssammudan.cotree.model.project.devposition.entity.ProjectDevPosition;
 import ssammudan.cotree.model.project.devposition.entity.QProjectDevPosition;
 import ssammudan.cotree.model.project.membership.entity.QProjectMembership;
 import ssammudan.cotree.model.project.project.entity.Project;
 import ssammudan.cotree.model.project.project.entity.QProject;
+import ssammudan.cotree.model.project.project.helper.ProjectQueryHelper;
 import ssammudan.cotree.model.project.techstack.entity.QProjectTechStack;
 
 /**
@@ -40,10 +37,12 @@ import ssammudan.cotree.model.project.techstack.entity.QProjectTechStack;
 public class ProjectRepositoryImpl implements ProjectRepositoryCustom {
 
 	private final JPAQueryFactory queryFactory;
+	private final ProjectQueryHelper projectQueryHelper;
 	private static final String SORT_BY_LIKE = "like";
 
-	public ProjectRepositoryImpl(EntityManager em) {
-		this.queryFactory = new JPAQueryFactory(em);
+	public ProjectRepositoryImpl(JPAQueryFactory queryFactory, ProjectQueryHelper projectQueryHelper) {
+		this.queryFactory = queryFactory;
+		this.projectQueryHelper = projectQueryHelper;
 	}
 
 	@Override
@@ -85,7 +84,7 @@ public class ProjectRepositoryImpl implements ProjectRepositoryCustom {
 			return new PageImpl<>(Collections.emptyList(), pageable, 0);
 
 		List<Project> projects = fetchProjectsWithDetails(projectIds);
-		List<ProjectListResponse> content = convertToDtoOrdered(projects, projectIds);
+		List<ProjectListResponse> content = projectQueryHelper.convertToDtoOrdered(projects, projectIds);
 
 		Long total = queryFactory.select(QProject.project.countDistinct())
 			.from(QProject.project)
@@ -102,13 +101,13 @@ public class ProjectRepositoryImpl implements ProjectRepositoryCustom {
 			return Collections.emptyList();
 
 		List<Project> projects = fetchProjectsWithDetails(projectIds);
-		return convertToDtoOrdered(projects, projectIds);
+		return projectQueryHelper.convertToDtoOrdered(projects, projectIds);
 	}
 
 	@Override
 	public Page<ProjectListResponse> findByFilters(Pageable pageable, List<Long> techStackIds,
 		List<Long> devPositionIds, String sort) {
-		BooleanBuilder where = buildFilterConditions(techStackIds, devPositionIds);
+		BooleanBuilder where = projectQueryHelper.buildFilterConditions(techStackIds, devPositionIds);
 
 		List<Long> filteredProjectIds = queryFactory
 			.select(QProject.project.id)
@@ -130,12 +129,12 @@ public class ProjectRepositoryImpl implements ProjectRepositoryCustom {
 		if (filteredProjectIds.isEmpty())
 			return new PageImpl<>(Collections.emptyList(), pageable, 0);
 
-		List<Long> sortedIds = sortFilteredProjects(filteredProjectIds, sort, pageable);
+		List<Long> sortedIds = projectQueryHelper.sortFilteredProjects(filteredProjectIds, sort, pageable);
 		if (sortedIds.isEmpty())
 			return new PageImpl<>(Collections.emptyList(), pageable, 0);
 
 		List<Project> projects = fetchProjectsWithDetails(sortedIds);
-		List<ProjectListResponse> content = convertToDtoOrdered(projects, sortedIds);
+		List<ProjectListResponse> content = projectQueryHelper.convertToDtoOrdered(projects, sortedIds);
 
 		return new PageImpl<>(content, pageable, filteredProjectIds.size());
 	}
@@ -177,73 +176,6 @@ public class ProjectRepositoryImpl implements ProjectRepositoryCustom {
 			.leftJoin(QProject.project.member).fetchJoin()
 			.where(QProject.project.id.in(ids))
 			.distinct()
-			.fetch();
-	}
-
-	private List<ProjectListResponse> convertToDtoOrdered(List<Project> projects, List<Long> orderedIds) {
-		Map<Long, ProjectListResponse> projectDtoMap = projects.stream()
-			.collect(Collectors.toMap(Project::getId, this::toDto));
-
-		return orderedIds.stream()
-			.map(projectDtoMap::get)
-			.toList();
-	}
-
-	private ProjectListResponse toDto(Project p) {
-		List<String> techStacksImageUrl = p.getProjectTechStacks().stream()
-			.map(ts -> ts.getTechStack().getImageUrl())
-			.toList();
-
-		long likeCount = p.getLikes().size();
-		int recruitmentCount = p.getProjectDevPositions().stream()
-			.mapToInt(ProjectDevPosition::getAmount).sum();
-
-		String description = p.getDescription();
-		if (description.length() > 30) {
-			description = description.substring(0, 30) + "...";
-		}
-
-		return new ProjectListResponse(
-			p.getId(),
-			p.getTitle(),
-			description,
-			p.getProjectImageUrl(),
-			p.getViewCount(),
-			likeCount,
-			recruitmentCount,
-			p.getIsOpen(),
-			p.getStartDate(),
-			p.getEndDate(),
-			techStacksImageUrl,
-			p.getMember().getUsername(),
-			p.getMember().getProfileImageUrl()
-		);
-	}
-
-	private BooleanBuilder buildFilterConditions(List<Long> techStackIds, List<Long> devPositionIds) {
-		BooleanBuilder where = new BooleanBuilder();
-		if (techStackIds != null && !techStackIds.isEmpty()) {
-			where.and(QProjectTechStack.projectTechStack.techStack.id.in(techStackIds));
-		}
-		if (devPositionIds != null && !devPositionIds.isEmpty()) {
-			where.and(QProjectDevPosition.projectDevPosition.developmentPosition.id.in(devPositionIds));
-		}
-		return where;
-	}
-
-	private List<Long> sortFilteredProjects(List<Long> ids, String sort, Pageable pageable) {
-		return queryFactory.select(QProject.project.id)
-			.from(QProject.project)
-			.leftJoin(QProject.project.likes, QLike.like)
-			.where(QProject.project.id.in(ids))
-			.groupBy(QProject.project.id)
-			.orderBy(
-				SORT_BY_LIKE.equalsIgnoreCase(sort)
-					? QLike.like.id.count().desc()
-					: QProject.project.createdAt.desc()
-			)
-			.offset(pageable.getOffset())
-			.limit(pageable.getPageSize())
 			.fetch();
 	}
 }
